fix: types exports
