+++ middlewares/referrer-policy/index.ts
@@ -1,19 +1,7 @@
 import { IncomingMessage, ServerResponse } from "http";
+import { ReferrerPolicyOptions, ReferrerPolicyToken } from "./types";
 
-type ReferrerPolicyToken =
-  | "no-referrer"
-  | "no-referrer-when-downgrade"
-  | "same-origin"
-  | "origin"
-  | "strict-origin"
-  | "origin-when-cross-origin"
-  | "strict-origin-when-cross-origin"
-  | "unsafe-url"
-  | "";
-
-export interface ReferrerPolicyOptions {
-  policy?: ReferrerPolicyToken | ReferrerPolicyToken[];
-}
+export * from "./types";
 
 const ALLOWED_TOKENS = new Set<ReferrerPolicyToken>([
   "no-referrer",
