+++ middlewares/referrer-policy/types.ts
@@ -1,0 +1,14 @@
+export type ReferrerPolicyToken =
+  | "no-referrer"
+  | "no-referrer-when-downgrade"
+  | "same-origin"
+  | "origin"
+  | "strict-origin"
+  | "origin-when-cross-origin"
+  | "strict-origin-when-cross-origin"
+  | "unsafe-url"
+  | "";
+
+export interface ReferrerPolicyOptions {
+  policy?: ReferrerPolicyToken | ReferrerPolicyToken[];
+}
