+++ middlewares/strict-transport-security/index.ts
@@ -1,12 +1,9 @@
 import { IncomingMessage, ServerResponse } from "http";
+import { StrictTransportSecurityOptions } from "./types";
+
+export * from "./types";
 
 const DEFAULT_MAX_AGE = 180 * 24 * 60 * 60;
-
-export interface StrictTransportSecurityOptions {
-  maxAge?: number;
-  includeSubDomains?: boolean;
-  preload?: boolean;
-}
 
 function parseMaxAge(value: number = DEFAULT_MAX_AGE): number {
   if (value >= 0 && Number.isFinite(value)) {
