+++ middlewares/strict-transport-security/types.ts
@@ -1,0 +1,5 @@
+export interface StrictTransportSecurityOptions {
+  maxAge?: number;
+  includeSubDomains?: boolean;
+  preload?: boolean;
+}
